requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pandas==2.3.3
numpy==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
python-multipart==0.0.22
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
threadpoolctl==3.6.0
six==1.17.0
pyyaml==6.0